renamed TourResources.js to StyledTourResources.js
