+++ src/pages/User/User.js
@@ -19,7 +19,7 @@ import AddIcon from '@material-ui/icons/Add'
 import CollectionDialog from '../../components/Collections/CollectionDialog.js'
 import { accountInfoSelector } from '../../redux/selectors'
 import CreateCollectionFab from '../../components/Miscellaneous/CreateCollectionFab.js'
-import { StyledTourResources } from '../../components/Miscellaneous/TourResources'
+import { StyledTourResources } from '../../components/Miscellaneous/StyledTourResources'
 import CollectionItem from '../../components/Collections/CollectionItem.js'
 import ShareTwitterDialog from '../../components/ShareTwitterDialog/ShareTwitterDialog.js'
 import { Link } from 'react-router-dom'
